requirements.txt
Appium-Python-Client==1.1.0
allure-pytest==2.8.40
pytest==6.1.1